+++ ext/curb.h
@@ -28,11 +28,11 @@
 #include "curb_macros.h"
 
 // These should be managed from the Rake 'release' task.
-#define CURB_VERSION   "1.0.2"
-#define CURB_VER_NUM   1002
+#define CURB_VERSION   "1.0.3"
+#define CURB_VER_NUM   1003
 #define CURB_VER_MAJ   1
 #define CURB_VER_MIN   0
-#define CURB_VER_MIC   2
+#define CURB_VER_MIC   3
 #define CURB_VER_PATCH 0