+++ ext/curb_multi.c
@@ -43,6 +43,11 @@ static void rb_curl_multi_remove(ruby_curl_multi *rbcm, VALUE easy);
 static void rb_curl_multi_read_info(VALUE self, CURLM *mptr);
 static void rb_curl_multi_run(VALUE self, CURLM *multi_handle, int *still_running);
 
+static VALUE callback_exception(VALUE did_raise, VALUE exception) {
+  rb_hash_aset(did_raise, rb_easy_hkey("error"), exception);
+  return exception;
+}
+
 void curl_multi_free(ruby_curl_multi *rbcm) {
   curl_multi_cleanup(rbcm->handle);
   free(rbcm);
@@ -292,13 +297,14 @@ static void rb_curl_mutli_handle_complete(VALUE self, CURL *easy_handle, int res
   }
 
   int status;
+  VALUE did_raise = rb_hash_new();
 
   if (!rb_easy_nil("complete_proc")) {
     callargs = rb_ary_new3(2, rb_easy_get("complete_proc"), easy);
     rbce->callback_active = 1;
-    rb_protect(call_status_handler1, callargs, &status);
+    rb_rescue(call_status_handler1, callargs, callback_exception, did_raise);
     rbce->callback_active = 0;
-    if (status) {
+    if (rb_hash_aref(did_raise, "error") != Qnil) {
       CURB_RB_CALLBACK_RAISE("complete")
     }
   }
@@ -316,9 +322,9 @@ static void rb_curl_mutli_handle_complete(VALUE self, CURL *easy_handle, int res
     if (!rb_easy_nil("failure_proc")) {
       callargs = rb_ary_new3(3, rb_easy_get("failure_proc"), easy, rb_curl_easy_error(result));
       rbce->callback_active = 1;
-      rb_protect(call_status_handler2, callargs, &status);
+      rb_rescue(call_status_handler2, callargs, callback_exception, did_raise);
       rbce->callback_active = 0;
-      if (status) {
+      if (rb_hash_aref(did_raise, "error") != Qnil) {
         CURB_RB_CALLBACK_RAISE("failure")
       }
     }
@@ -327,17 +333,17 @@ static void rb_curl_mutli_handle_complete(VALUE self, CURL *easy_handle, int res
     /* NOTE: we allow response_code == 0, in the case of non http requests e.g. reading from disk */
     callargs = rb_ary_new3(2, rb_easy_get("success_proc"), easy);
     rbce->callback_active = 1;
-    rb_protect(call_status_handler1, callargs, &status);
+    rb_rescue(call_status_handler1, callargs, callback_exception, did_raise);
     rbce->callback_active = 0;
-    if (status) {
+    if (rb_hash_aref(did_raise, "error") != Qnil) {
       CURB_RB_CALLBACK_RAISE("success")
     }
   } else if (!rb_easy_nil("redirect_proc") && ((response_code >= 300 && response_code < 400) || redirect_count > 0) ) {
     rbce->callback_active = 1;
     callargs = rb_ary_new3(3, rb_easy_get("redirect_proc"), easy, rb_curl_easy_error(result));
     rbce->callback_active = 0;
-    rb_protect(call_status_handler2, callargs, &status);
-    if (status) {
+    rb_rescue(call_status_handler2, callargs, callback_exception, did_raise);
+    if (rb_hash_aref(did_raise, "error") != Qnil) {
       CURB_RB_CALLBACK_RAISE("redirect")
     }
   } else if (!rb_easy_nil("missing_proc") &&
@@ -345,17 +351,17 @@ static void rb_curl_mutli_handle_complete(VALUE self, CURL *easy_handle, int res
     rbce->callback_active = 1;
     callargs = rb_ary_new3(3, rb_easy_get("missing_proc"), easy, rb_curl_easy_error(result));
     rbce->callback_active = 0;
-    rb_protect(call_status_handler2, callargs, &status);
-    if (status) {
+    rb_rescue(call_status_handler2, callargs, callback_exception, did_raise);
+    if (rb_hash_aref(did_raise, "error") != Qnil) {
       CURB_RB_CALLBACK_RAISE("missing")
     }
   } else if (!rb_easy_nil("failure_proc") &&
           (response_code >= 500 && response_code <= 999)) {
     callargs = rb_ary_new3(3, rb_easy_get("failure_proc"), easy, rb_curl_easy_error(result));
     rbce->callback_active = 1;
-    rb_protect(call_status_handler2, callargs, &status);
+    rb_rescue(call_status_handler2, callargs, callback_exception, did_raise);
     rbce->callback_active = 0;
-    if (status) {
+    if (rb_hash_aref(did_raise, "error") != Qnil) {
       CURB_RB_CALLBACK_RAISE("failure")
     }
   }