+++ ChangeLog.md
@@ -1,4 +1,6 @@
 # Change Log
+## 1.0.3
+* Adjust how we handle exceptions to handle more cases
 ## 1.0.2
 * Address alloc function T_DATA see: https://bugs.ruby-lang.org/issues/18007
 * Add CURLOPT_PROXY_SSL_VERIFYHOST support